Move methods and state from Movies to Main
Rename JS, CSS, test files from Movies to CardContainer

Update testing in Main to reflect class component

Rename files from Movies to CardContainer
